refactor: rename translation schema to schemas directory to be synced
