+++ src/core/translation/locales/en/schemas/schema.json
