+++ src/core/translation/locales/jp/schemas/schema.json
