+++ src/core/translation/locales/jp/translation.ts
@@ -1,8 +1,8 @@
 import fileUploader from './fileUploader.json';
 import forgetPassword from './forgetPassword.json';
 import orgProfile from './organizations/organizations.json';
-import schema from './schema/schema.json';
-import schemaCreate from './schema/schemaCreate.json';
+import schema from './schemas/schema.json';
+import schemaCreate from './schemas/schemaCreate.json';
 import settings from './settings/settings.json';
 import signIn from './signIn.json';
 import detail from './singUp/detail.json';
